WIP: add search bar and functionality to refresh list
