+++ src/components/AdminInactiveUsers/AdminInactiveUsers.css
@@ -1,0 +1,89 @@
+.TableHeader {
+    display: flex;
+    justify-content: space-between;
+    padding: 1rem 1rem 0 0;
+}
+
+.SectionTitle {
+    display: flex !important;
+    justify-content: space-between !important;
+}
+
+.SearchArea,
+.SectionTitle {
+    display: grid;
+    align-items: center;
+    padding-top: 1rem;
+}
+
+.RefreshButton:hover {
+    cursor: pointer;
+    display: inline;
+}
+
+.SpinnerArea {
+    align-items: center;
+    justify-items: center;
+}
+
+/* ////////////---- calendar styles ---- ///////////// */
+.DateSection {
+    padding: 2px 0px 2px 0px;
+    display: flex;
+    flex-direction: row;
+    font-size: 15px;
+    justify-content: space-between;
+    align-items: center;
+    height: 25px;
+    gap: 0.5rem;
+    margin-top: 3px;
+}
+
+.DateItem {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.DateSection:hover {
+    cursor: pointer;
+}
+
+.CalenderToposition,
+.CalenderFromposition {
+    background-color: #ffffff;
+    border: 1px solid #cccccc;
+    border-radius: 5px;
+    padding: 0.5rem;
+    display: grid;
+    gap: 1rem;
+    position: absolute;
+    z-index: 10;
+    left: 0.1rem;
+    width: 58%;
+}
+
+.dateField {
+    width: 7rem;
+}
+
+.OuterFilterItem {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    background-color: #fff;
+    width: 20rem;
+    height: 40px;
+    border-radius: 9px;
+    border: 1px solid #000000;
+}
+
+.CalendarSection {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding-top: 1rem;
+    position: relative;
+    float: right;
+}
